Rename from sknnr-spatial to sklearn-raster
