+++ pyproject.toml
@@ -3,9 +3,9 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "sknnr-spatial"
+name = "sklearn-raster"
 dynamic = ["version"]
-description = "Spatial raster prediction with scikit-learn kNN estimators."
+description = "Spatial raster prediction with scikit-learn estimators."
 readme = "README.md"
 requires-python = ">=3.9"
 authors = [
@@ -29,21 +29,21 @@ datasets = [
     "pooch",
 ]
 tutorials = [
-    "sknnr-spatial[datasets]",
+    "sklearn-raster[datasets]",
     "matplotlib",
 ]
 
 [project.urls]
-Homepage = "https://github.com/lemma-osu/sknnr-spatial"
+Homepage = "https://github.com/lemma-osu/sklearn-raster"
 
 [tool.hatch.version]
-path = "src/sknnr_spatial/__init__.py"
+path = "src/sklearn_raster/__init__.py"
 
 [tool.hatch.build]
-exclude = ["src/sknnr_spatial/datasets/data/"]
+exclude = ["src/sklearn_raster/datasets/data/"]
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/sknnr_spatial"]
+packages = ["src/sklearn_raster"]
 
 [tool.hatch.build.targets.sdist]
 include = ["/src"]
@@ -52,19 +52,19 @@ include = ["/src"]
 dependencies = [
     "pre-commit",
     "ipykernel",
-    "sknnr-spatial[datasets]",
+    "sklearn-raster[datasets]",
 ]
 
 [tool.hatch.envs.test]
 dependencies = [
     "pytest",
     "pytest-cov",
-    "sknnr-spatial[datasets]",
+    "sklearn-raster[datasets]",
 ]
 
 [tool.hatch.envs.test.scripts]
 all = "pytest . {args} --doctest-modules"
-cov = "pytest . --cov=src/sknnr_spatial {args} --doctest-modules"
+cov = "pytest . --cov=src/sklearn_raster {args} --doctest-modules"
 
 [tool.hatch.envs.test_matrix]
 template = "test"
@@ -89,7 +89,7 @@ show-fixes = true
 
 [tool.ruff.lint]
 select = ["E", "I", "F", "B", "FA", "UP", "PT", "Q", "RET", "SIM", "PERF"]
-isort.known-first-party = ["sknnr_spatial"]
+isort.known-first-party = ["sklearn_raster"]
 
 [tool.hatch.metadata]
 allow-direct-references = true
