+++ src/sklearn_raster/datasets/_registry.py
