+++ src/sklearn_raster/datasets/data/swo_ecoplot_128x128.zip
