+++ src/sklearn_raster/types.py
