+++ src/sklearn_raster/utils/estimator.py
