+++ src/sklearn_raster/utils/image.py
